Move AndroidX lifecycle coroutineScope extensions
…to LifecycleDefaultScopes.kt

That is to prevent class initialization in case they are not used.
