+++ integration/kotlinx-coroutines-androidx-lifecycle/src/Lifecycle.kt
@@ -8,31 +8,6 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.SupervisorJob
-import java.util.concurrent.ConcurrentHashMap
-
-/**
- * Returns a [CoroutineScope] that uses [Dispatchers.Main] by default, and that is cancelled when
- * the [Lifecycle] reaches [Lifecycle.State.DESTROYED] state.
- *
- * Note that this value is cached until the Lifecycle reaches the destroyed state.
- */
-val Lifecycle.coroutineScope: CoroutineScope
-    get() = lifecycleCoroutineScopes[this] ?: job.let { job ->
-        val newScope = CoroutineScope(job + Dispatchers.Main)
-        if (job.isActive) {
-            lifecycleCoroutineScopes[this] = newScope
-            job.invokeOnCompletion { _ -> lifecycleCoroutineScopes -= this }
-        }
-        newScope
-    }
-
-/**
- * Calls [Lifecycle.coroutineScope] for the [Lifecycle] of this [LifecycleOwner].
- *
- * This is an inline property, just there for convenient usage from any [LifecycleOwner],
- * like FragmentActivity, AppCompatActivity, Fragment and LifecycleService.
- */
-inline val LifecycleOwner.coroutineScope get() = lifecycle.coroutineScope
 
 /**
  * Returns a [CoroutineScope] that uses [Dispatchers.Main] by default, and that will be cancelled as
@@ -45,22 +20,6 @@ inline val LifecycleOwner.coroutineScope get() = lifecycle.coroutineScope
 fun Lifecycle.createScope(activeWhile: Lifecycle.State): CoroutineScope {
     return CoroutineScope(createJob(activeWhile) + Dispatchers.Main)
 }
-
-/**
- * Returns a [SupervisorJob] that will be cancelled as soon as the [Lifecycle] reaches
- * [Lifecycle.State.DESTROYED] state.
- *
- * Note that this value is cached until the Lifecycle reaches the destroyed state.
- *
- * You can use this job for custom [CoroutineScope]s, or as a parent [Job].
- */
-val Lifecycle.job: Job
-    get() = lifecycleJobs[this] ?: createJob().also {
-        if (it.isActive) {
-            lifecycleJobs[this] = it
-            it.invokeOnCompletion { _ -> lifecycleJobs -= this }
-        }
-    }
 
 /**
  * Creates a [SupervisorJob] that will be cancelled as soon as this [Lifecycle]
@@ -86,6 +45,3 @@ fun Lifecycle.createJob(activeWhile: Lifecycle.State = INITIALIZED): Job {
         })
     }
 }
-
-private val lifecycleJobs = ConcurrentHashMap<Lifecycle, Job>()
-private val lifecycleCoroutineScopes = ConcurrentHashMap<Lifecycle, CoroutineScope>()
